解决1、角色或部门不再需要认领了，根据流程模型设置情况来实际审批 2、支持角色、部门到下级全体人员审批，也支持原有单个人审批逻辑，都做兼容

+++ shengyu-module-system/shengyu-module-system-biz/src/main/java/com/shengyu/module/system/controller/admin/flow/ProcessTaskController.java
@@ -196,7 +196,7 @@ public class ProcessTaskController {
     @Operation(summary = "当前用户待办任务数量")
     @PreAuthorize("@ss.hasPermission('flw:processTask:countPendingApproval')")
     @PostMapping("/count-pending-approval")
-    public CommonResult<Integer> countPendingApproval() {
+    public CommonResult<Long> countPendingApproval() {
         return success(processTaskService.countPendingApproval());
     }
